# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jewish prayer times calculation and publisher management",
        "title": "Zmanim Lab API",
        "termsOfService": "https://zmanim-lab.com/terms",
        "contact": {
            "name": "Zmanim Lab Support",
            "email": "support@zmanim-lab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cities": {
            "get": {
                "description": "Search for cities by name with optional filtering by country, region, or continent. Uses fuzzy matching for typo tolerance.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Search cities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (min 2 chars for fuzzy match)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO 3166-1 alpha-2 country code",
                        "name": "country_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Region/state name",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Continent code (AF, AS, EU, NA, OC, SA, AN)",
                        "name": "continent_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 10, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching cities",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_models.CitySearchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/cities/nearby": {
            "get": {
                "description": "Finds the nearest city to the given GPS coordinates using PostGIS spatial queries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Find nearest city",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude (-90 to 90)",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude (-180 to 180)",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nearest city with distance",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid coordinates",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handlers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/internal_handlers.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jcom-dev_zmanim-lab_internal_models.City": {
            "type": "object",
            "properties": {
                "continent": {
                    "description": "Continent code (AF, AN, AS, EU, NA, OC, SA)",
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "display_name": {
                    "description": "Computed display field",
                    "type": "string"
                },
                "elevation": {
                    "description": "Elevation in meters above sea level",
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "population": {
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "region_type": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "github_com_jcom-dev_zmanim-lab_internal_models.CitySearchResponse": {
            "type": "object",
            "properties": {
                "cities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jcom-dev_zmanim-lab_internal_models.City"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handlers.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handlers.ResponseMeta"
                }
            }
        },
...
        "internal_handlers.ResponseMeta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token from Clerk authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}